remove inner action listener and move it to BankController
